support java-library plugin and add test for dependency recommender
the new java-library adds new configurations for users (api and
implenentation). these new configurations were causing the
MavenResolvedDependenciesPlugin to be unable to set the version
correctly in the generated POM file.

the fix is to pull the dependencies from the compileClasspath and
runtimeClasspath configurations and let the resolve logic look in the
proper scope.