Remove MavenCompileOnlyPlugin
This plugin takes compileOnly dependencies and adds them to the generated POM
with provided scope. This sometimes causes issues with projects by generating
two distinct dependencies blocks.

By removing it we better align ourselves with Gradle's position, which is that
compileOnly dependencies do not need to be published in any scope[1].

[1]: https://discuss.gradle.org/t/publishing-plugin-should-respect-compileonly-configuration/22903/2